requirements.txt
kafka-python
apache-airflow[elasticsearch]
